Moved booking logic to calcom/features
Needed to reuse on public API
